Add more extension points to the k8s API.
In preparation for making use of Kubernetes' ThirdPartyResource for
storage of DTabs, I have made our existing k8s API integration library
more generic and flexible.

Note that some of the type signatures are a little inelegant; this is
required because Jackson polymorphism and generics don'at always
interact well, so some things that don't technically need to be
type-specialized (Watches and Lists) are.